pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aituber-starter-kit"
version = "0.1.0"
description = "Open-source toolkit for creating AI-powered VTubers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AITuber Community"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["aituber", "vtuber", "ai", "streaming", "live2d", "vtube-studio"]
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pytchat>=0.5.5",
    "openai>=1.50.0",
    "pyvts>=0.3.3",
    "sounddevice>=0.4.7",
    "numpy>=1.26.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Dashboard
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
aituber = "src.main:run"

[project.urls]
Homepage = "https://github.com/aituber-community/aituber-starter-kit"
Documentation = "https://github.com/aituber-community/aituber-starter-kit#readme"
Repository = "https://github.com/aituber-community/aituber-starter-kit"
Issues = "https://github.com/aituber-community/aituber-starter-kit/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
